dedicated component for working hours view
