+++ scripts/extensions/availability-manager/src/interfaces.ts
@@ -1,0 +1,14 @@
+import {IBaseRestApiResponse} from 'superdesk-api';
+
+export interface IWorkingHours {
+    start_time: string;
+    end_time: string;
+    tags: Array<{name: string; code: string}>;
+}
+
+export interface IAvailabilityRecord extends IBaseRestApiResponse {
+    date: string;
+    status: 'available' | 'unavailable' | 'partial';
+    working_hours?: Array<IWorkingHours>;
+    language: string;
+}
