--- scripts/extensions/availability-manager/src/settings/day-view-edit.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-interface IProps {
-
-}
-
-export class DayViewEdit extends React.PureComponent<IProps> {
-    render() {
-        return null
-    }
-}
